fix markdown links
markdown link syntax is `[text](href)` I never forgot again after someone pointed out that the bits at the edge of the square bracket are like the link underline that you click on. Also, it seems that github markdown does not work for markdown links, but it does work like this.